services/cluster-services/frame-db/storage_services/ingestion_worker/service/requirements.txt
python-json-logger
kubernetes
redis
sqlalchemy
PyMySQL
sqlalchemy_utils
kafka-python